+++ libs/cardano-ledger-core/app/CLI.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE NumericUnderscores #-}
+
 module CLI (
   Opts (..),
   optsParser,
@@ -6,9 +8,11 @@ module CLI (
 import Cardano.Ledger.Binary (mkVersion64)
 import Cardano.Ledger.Plutus.Evaluate
 import Options.Applicative
+import Text.Read (readMaybe)
 
 data Opts = Opts
   { optsScriptWithContext :: !String
+  , optsTimeout :: !Int
   , optsOverrides :: !PlutusDebugOverrides
   }
   deriving (Show)
@@ -19,6 +23,7 @@ overridesParser =
     <$> option
       (Just <$> str)
       ( long "script"
+          <> short 's'
           <> value Nothing
           <> help "Plutus script hex without context"
       )
@@ -32,12 +37,14 @@ overridesParser =
     <*> option
       (Just <$> auto)
       ( long "language"
+          <> short 'l'
           <> value Nothing
           <> help "Plutus language version"
       )
     <*> option
-      (str >>= pure . Just . map read . words)
+      (mapM readMaybe . words <$> str)
       ( long "cost-model-values"
+          <> short 'c'
           <> value Nothing
           <> help ""
       )
@@ -53,10 +60,28 @@ overridesParser =
           <> value Nothing
           <> help ""
       )
+    <*> switch
+      ( long "enforce-execution-units"
+          <> help
+            ( "By default plutus-debug upon a failure will re-evaluate supplied script one more time "
+                <> "without bounding execution in order to report expected execution units. "
+                <> "In case when this unbounded computation is a problem, this flag allows for "
+                <> "disabling this reporting of expected execution units."
+            )
+      )
 
 optsParser :: Parser Opts
 optsParser =
   Opts
-    <$> strArgument
-      (metavar "SCRIPT_WITH_CONTEXT(BASE64)")
+    <$> strArgument (metavar "SCRIPT_WITH_CONTEXT(BASE64)")
+    <*> option
+      auto
+      ( long "timeout"
+          <> short 't'
+          <> value 5_000_000
+          <> help
+            ( "Timeout in number of milliseconds. Default is 5000000 ms (or 5 seconds). "
+                <> "Specifying a negative number will effectively remove the timeout and unbound execution."
+            )
+      )
     <*> overridesParser