+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 1.18.0.0
 
+* Add a timeout argument to `plutusDebug`
+* Add `pdoExUnitsEnforced` to `PlutusDebugOverrides` and add `defaultPlutusDebugOverrides`
+* Add `NFData` instance for `PlutusDebugInfo`
+* Add `DebugTimedOut` constructor to `PlutusDebugInfo`
+* Add `debugPlutusUnbounded`
 * Added `binaryUpgradeTx`, `binaryUpgradeTxBody`, `binaryUpgradeTxWits`, `binaryUpgradeTxAuxData`
 * Remove `upgradeTx` and `TxUpgradeError` from `EraTx`
 * Remove `upgradeTxBody` and `TxBodyUpgradeError` from `EraTxBody`